refactor: split story
